requirements.txt
git+https://github.com/django/django.git#egg=django
psycopg2
djangorestframework
markdown
django-filter
djangorestframework-gis
